+++ waypoints/src/main/kotlin/de/md5lukas/waypoints/gui/pages/SelectBeaconColorPage.kt
@@ -9,7 +9,6 @@ import de.md5lukas.kinvs.items.GUIContent
 import de.md5lukas.kinvs.items.GUIItem
 import de.md5lukas.waypoints.api.Waypoint
 import de.md5lukas.waypoints.gui.WaypointsGUI
-import de.md5lukas.waypoints.pointers.BeaconColor
 import net.kyori.adventure.text.Component
 import org.bukkit.inventory.ItemStack
 
@@ -44,6 +43,7 @@ class SelectBeaconColorPage(wpGUI: WaypointsGUI, private val waypoint: Waypoint)
                   switchContext(SynchronizationContext.SYNC)
                   wpGUI.playSound { clickSuccess }
                   wpGUI.goBack()
+                  (wpGUI.gui.activePage as BasePage).update()
                 }
               })
         }